Put model to a file
